+++ internal/components/components_test.go
@@ -55,6 +55,7 @@ func TestDefaultComponents(t *testing.T) {
 		"kafkametrics",
 		"kubeletstats",
 		"mongodbatlas",
+		"oracledb",
 		"otlp",
 		"prometheus",
 		"prometheus_exec",